+++ package.json
@@ -77,6 +77,7 @@
     "short-number": "^1.0.7",
     "slick-carousel": "^1.8.1",
     "tailwind-merge": "^2.5.4",
+    "tailwindcss-animate": "^1.0.7",
     "ual-anchor": "^1.3.0",
     "ual-reactjs-renderer": "^0.3.1",
     "universal-authenticator-library": "^0.3.0",
